refactor: extract HTTP transport policy
